Split ReadArgumentCheckNode inner classes into their own files
